requirements.txt
fastapi
langchain
langchain_core
langchain-community
uvicorn[standard]
python-dotenv
tqdm
tiktoken
langchain-openai
faiss-cpu
langgraph
crewai
wikipedia
tavily-python
google-search-results